feat(crawl-redis): reduce `crawl:<id>:visited` size in Redis by 16x
Will help with memory exhaustion and eviction of other important keys.

The profiling of our eviction Redis revealed that `crawl:<id>:visited` keys are using 30GB in total. This is due to how we do URL deduplication. Whenever `lockURL` (or one of its derivatives) is called, and `deduplicateSimilarURLs` is enabled in `crawlerOptions` (default), it will run `generateURLPermutations` to generate up to 16 permutations of the URL, and insert that into the visited key. If not all generated permutations were inserted, that means that the URL was already visited by the crawl. However, this means that one URL is stored 16 times as various versions, which bloats memory usage.

The `generateURLPermutations` function transforms a URL into 16 slightly different versions of itself. However, due to the implementation, it actually transforms any of the permutations into the same 16 versions as both the source URL and any other URL from the same permutations. We can use this to only store the first generated URL in Redis and achieve the same functionality.

This is also backwards-compatible and non-disruptive, since the previous code version pushed all permutations to the set, and we are now checking for the first permutation, which is surely in there.

This statement about the implementation is mostly proven by `permu-refactor.test.ts` (also see the comment in `crawl-redis.ts` for more details).